+++ .eslintrc.js
@@ -2,7 +2,6 @@ module.exports = {
 	root: true,
 	extends: [
 		'mitmaro',
-		'mitmaro/config/ecmascript-8',
-		'mitmaro/config/node',
+		'mitmaro/config/typescript',
 	],
 };