+++ scripts/build.bash
@@ -2,6 +2,7 @@
 
 cd "$(dirname "${BASH_SOURCE[0]}")" && source "./common.bash"
 
-babel src/ -d lib/
+rm -rf "./lib/"
+rm -rf "./types/"
 
-cp src/dependency-manager.d.ts lib/
+tsc -p tsconfig-build.json "$@"