+++ scripts/lint.bash
@@ -3,7 +3,7 @@
 cd "$(dirname "${BASH_SOURCE[0]}")" && source "./common.bash"
 
 info "Running ESLint"
-eslint "$@" .
+eslint --ext .ts --fix "$@" .
 
 info "Running TSLint"
-tslint --project tsconfig.json --format verbose --config tslint.json "$@" 'src/**/*.ts'
+tslint --project tsconfig.json --format stylish --config tslint.json "$@" 'src/**/*.ts'