+++ test/.eslintrc.js
@@ -2,9 +2,12 @@ module.exports = {
 	root: true,
 	extends: [
 		'mitmaro',
-		'mitmaro/config/ecmascript-8',
-		'mitmaro/config/node',
 		'mitmaro/config/mocha',
 		'mitmaro/config/chai',
+		'mitmaro/config/typescript-mocha',
 	],
+	rules: {
+		'security/detect-non-literal-require': 'off',
+		'no-loop-func': 'off',
+	}
 };