+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 The format conforms to [Keep a Changelog](http://keepachangelog.com/).
 
+## 0.2.0 - 2018-11-11
+### Added
+- Callback function for when dependency is loaded
+
+### Changed
+- Rewritten in TypeScript
+
 ## 0.1.0 - 2018-02-11
 ### Added
 - Initial project release